+++ EAP/src/main/java/de/rub/nds/tlsattacker/eap/FragState.java
@@ -74,7 +74,7 @@ public class FragState implements EapState {
     @Override
     public byte[] receive() {
         data = nic.receiveFrame();
-        id = (int) data[19]; // Get ID
+        id = data[19]; // Get ID
 
         LOGGER.debug("receive() TLS-FLAG: {}", Byte.toString(data[23]));
