+++ EAP/src/main/java/de/rub/nds/tlsattacker/eap/NoFragState.java
@@ -58,7 +58,7 @@ public class NoFragState implements EapState {
     @Override
     public byte[] receive() {
         data = nic.receiveFrame();
-        id = (int) data[19]; // Get ID
+        id = data[19]; // Get ID
 
         LOGGER.debug("receive() TLS-FLAG: {}", Byte.toString(data[23]));
